chore: reorganize rule types declaration
