+++ src/functions/enumeration.ts
@@ -1,4 +1,10 @@
-import { IEnumRuleOptions, IFunction, IFunctionResult } from '../types';
+import { IFunction, IFunctionResult, IRule, RuleFunction } from '../types';
+
+export interface IEnumRuleOptions {
+  values: Array<string | number>;
+}
+
+export type EnumRule = IRule<RuleFunction.ENUM, IEnumRuleOptions>;
 
 export const enumeration: IFunction<IEnumRuleOptions> = (targetVal, opts) => {
   const results: IFunctionResult[] = [];
