+++ src/functions/length.ts
@@ -1,4 +1,11 @@
-import { IFunction, IFunctionResult, ILengthRuleOptions } from '../types';
+import { IFunction, IFunctionResult, IRule, RuleFunction } from '../types';
+
+export interface ILengthRuleOptions {
+  min?: number;
+  max?: number;
+}
+
+export type LengthRule = IRule<RuleFunction.LENGTH, ILengthRuleOptions>;
 
 export const length: IFunction<ILengthRuleOptions> = (targetVal, opts) => {
   const results: IFunctionResult[] = [];
